renaming and standardizing stuff
