+++ src/components/about/About.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './about.css';
+import './About.css';
 
 export default class About extends React.Component {
     render() {
