+++ src/components/app/App.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Splash from './splash/splash.jsx';
-import Maggie from './maggie/maggie.jsx';
-import About from './about/about.jsx';
-import './index.css';
+import Splash from '../splash/Splash.jsx';
+import Maggie from '../maggie/Maggie.jsx';
+import About from '../about/About.jsx';
 import { HashRouter, Route, Switch } from 'react-router-dom';
