+++ src/components/splash/pithy/PithyText.jsx
